Updated package.json and README.md
