+++ README.md
@@ -3,43 +3,40 @@ Lorem fill is a tool for fast prototyping of web pages by programmtically fillin
 
 Syntax is super easy and follows conventions from other frameworks.
 
-Add the attribute lorem-fill="x" with x being any number. LoremFill will repeat the element x times only on page load and won't add more to your raw HTML file.
+Add the attribute lorem-pg="x" with x being any number. LoremFill will repeat the element x times only on page load and won't add more to your raw HTML file.
 
 ```HTML
-    <p lorem-fill="5" class="test"></p>
+    <p lorem-pg="5" class="test"></p>
 ```
 
 
-
-
 ```HTML
-    <p lorem-fill="5" class="test"></p>
-    <p lorem-fill="2" class="test">{{lorem-pg}}</p>
+    <p lorem-pg="5" class="test"></p>
+    <h1 lorem-sent="2" class="test"></h1>
 ```
 
-Above you'll notice a '{{lorem-pg}}', this markup programmatically adds a paragraph's worth of text. Review the following table for more information:
+ To determine the output of the attribute simply review the following table:
 
 
 | Markup        | output        |
 | ------------- |:-------------:| 
-| {{lorem}}     | Lorem         |
-| {{lorem-sent}}| Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.      |
-| {{lorem-pg}}  | Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.      |
-| {{lorem-list}}| "Lorem, ipsum, dolor, sit, amet, consectetur, adipiscing, elit, sed, do, eiusmod, tempor, incididunt, ut, labore, et, dolore, magna, aliqua |
+| lorem="x"     | Will output a random filler word |
+| lorem-sent="x"| Creates a 8 word sentence |
+| lorem-pg="x"  | Creates a 10 sentence paragraph |
 
 
-By default any element without text will yield nothing. Conversely, if it has text but lacks the mark up such as the following: 
+By default any element if it has text it will yield the text entered. 
 
 ```HTML
-    <p lorem-fill="2" class="test">This is some awesome text</p>
+    <p lorem-pg="2" class="test">This is some awesome text</p>
 ```
 
-This will ignore the lorem-fill attribute and only output one paragraph with the text "This is some awesome text".
+This will ignore the lorem-pg attribute and only output one paragraph with the text "This is some awesome text".
 
-Any cloned element created by a lorem-fill will carry the attributes of their parent with the exception of the lorem-fill attribute. Hence the following will be the output:
+Any cloned element created by a lorem attribute will carry the attributes of their parent with the exception of the lorem lorem attribute. Hence the following will be the output:
 
 ```HTML
-    <p lorem-fill="2" class="test">{{lorem-sent}}</p>
+    <p lorem-sent="2" class="test"></p>
     <p class="test">
         Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.
     </p>
@@ -48,6 +45,41 @@ Any cloned element created by a lorem-fill will carry the attributes of their pa
     </p>
 
 ```
+
+In addition, nested objects without text will essentially create a template for cloning. Such as the following: 
+
+```HTML
+    <ul>
+        <li lorem="6">
+            <a href="#"></a>
+        </li>
+    </ul>
+```
+
+This would yield 6 list tags each with a nested anchor. The anchor tag will have the outputed text. 
+
+```HTML
+    <ul>
+        <li lorem="6">
+            <a href="#">lorem</a>
+        </li>
+        <li>
+            <a href="#">dolor</a>
+        </li>
+        <li>
+            <a href="#">ipsum</a>
+        </li>
+        <li>
+            <a href="#">pariatur</a>
+        </li>
+         <li>
+            <a href="#">cupidatat</a>
+        </li>
+        <li>
+            <a href="#">irure</a>
+        </li>
+    </ul>
+``` 
 
 Feel free to test it out and let me know of any issues! bobby@robertorivera.dev
 Also, if you're feeling generous and want to feed my coffee habit consider clicking the link below:
